+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - webapp: Improve parallel fetch of database
 
+### Fixed
+
+- webapp: Fix updates of year view while loading
+
 ## [1.6.1] - 2022-08-13
 
 ### Fixed